MAINT: move silhouette_score
